NXDRIVE-903: Clean-up
* Fixed a subtle threading bug with time.strptime()
* Fixed a RemoteFileInfo instanciation bug into RemoteWatcher._update_remote_states()
* Added Engine.__repr__()
* Better repr(FileInfo)
* Remove unused LocalClient.is_inside()
* Vars clean-up in various Processor methods
* Vars clean-up in RemoteWatcher._update_remote_states()
* Refactor Tracker events sending
* Use os.path.is*() instead of general os.path.exists()
* Keep the testing database in Engine.unbind()
* Replace use of dict.iteritems() with dict.items() to ease future Python 3 transition
* Give time for test_nxdrive_903.py to perform renames

+++ nxdrive/engine/engine.py
@@ -67,7 +67,7 @@ class Engine(QObject):
     newReadonly = pyqtSignal(object, object)
     deleteReadonly = pyqtSignal(object)
     newLocked = pyqtSignal(object, object, object)
-    newSync = pyqtSignal(object, object)
+    newSync = pyqtSignal(object)
     newError = pyqtSignal(object)
     newQueueItem = pyqtSignal(object)
     offline = pyqtSignal()
@@ -156,6 +156,10 @@ class Engine(QObject):
         # Pause in case of no more space on the device
         self.noSpaceLeftOnDevice.connect(self.suspend)
 
+    def __repr__(self):
+        fmt = '{name}<name={cls.name!r}, uid={cls.uid!r}, type={cls.type!r}>'
+        return fmt.format(name=type(self).__name__, cls=self)
+
     @pyqtSlot(object)
     def _check_sync_start(self, row_id):
         if not self._sync_started:
@@ -174,14 +178,13 @@ class Engine(QObject):
             self.start()
 
     def stop_processor_on(self, path):
-        for worker in self.get_queue_manager().get_processors_on(
-                path, exact_match=True):
+        for worker in self.get_queue_manager().get_processors_on(path):
             log.trace('Quitting processor: %r as requested to stop on %r',
                       worker, path)
             worker.quit()
 
     def set_local_folder(self, path):
-        log.debug("Update local folder to '%s'", path)
+        log.debug('Update local folder to %r', path)
         self.local_folder = path
         self._local_watcher.stop()
         self._create_local_watcher()
@@ -356,13 +359,15 @@ class Engine(QObject):
 
         self.dispose_db()
         log.debug('Remove DB file %r', self._get_db_file())
-        try:
-            os.remove(self._get_db_file())
-        except (IOError, OSError) as exc:
-            if exc.errno != 2:  # File not found, already removed
-                log.exception('Database removal error')
-
         self._manager.osi.unregister_folder_link(self.local_folder)
+
+        # Keep the database for tests
+        if not os.environ.get('WORKSPACE', False):
+            try:
+                os.remove(self._get_db_file())
+            except (IOError, OSError) as exc:
+                if exc.errno != 2:  # File not found, already removed
+                    log.exception('Database removal error')
 
     def check_fs_marker(self):
         tag = 'drive-fs-test'
@@ -477,9 +482,10 @@ class Engine(QObject):
     def create_thread(self, worker=None, name=None, start_connect=True):
         if worker is None:
             worker = Worker(self, name=name)
-        # If subclass of Processor then connect the newSync signal
+
         if isinstance(worker, Processor):
             worker.pairSync.connect(self.newSync)
+
         thread = worker.get_thread()
         if start_connect:
             thread.started.connect(worker.run)
